@@ -13,15 +13,16 @@ Gem::Specification.new do |s|
   s.description = "Content Manager for RoR"
 
   s.rubyforge_project = "my_admin"
+  s.required_ruby_version = ">= 1.9.3"
 
-  s.add_dependency "rails",           "~> 4.1.0"
-  s.add_dependency "breadcrumbs"  ,   "~> 0.1.6"
-  s.add_dependency "dynamic_form" ,   "~> 1.1.4"
-  s.add_dependency "paperclip" ,      "~> 4.2.0"
-  s.add_dependency "will_paginate",   "~> 3.0.7"
-  s.add_dependency "spreadsheet",     "~> 1.0.0"
-  s.add_dependency "htmlentities",    "~> 4.3.2"
-  s.add_dependency "ckeditor",        "~> 4.1.1"
+  s.add_dependency "rails",           "~> 4.2.1"
+  s.add_dependency "breadcrumbs"  ,   "~> 0.1"
+  s.add_dependency "dynamic_form" ,   "~> 1.1"
+  s.add_dependency "paperclip" ,      "~> 4.2"
+  s.add_dependency "will_paginate",   "~> 3.0"
+  s.add_dependency "spreadsheet",     "~> 1.0"
+  s.add_dependency "htmlentities",    "~> 4.3"
+  s.add_dependency "ckeditor",        "~> 4.1"
 
   s.files         = `git ls-files`.split("\n")
   s.test_files    = `git ls-files -- {test,spec,features}/*`.split("\n")
