@@ -12,7 +12,7 @@ module MyAdmin
         MyAdmin::Application.items.each do |app|
           app.models.each do |model|
             model.my_admin.permissions.each do |permission|
-              MyAdmin::Permission.find_or_create_by_application_and_model_and_name({:application => app.key, :model => model.to_s, :name => permission})
+              MyAdmin::Permission.find_or_create_by(application: app.key, model: model.to_s, name: permission)
             end
           end
         end
