@@ -1,3 +1,3 @@
 module MyAdmin
-  VERSION = "0.0.2"
+  VERSION = "0.0.7"
 end
